\name{conditionals}
\alias{ifo}
\alias{ife}
\alias{ifn}
\title{conditional utilities}
\description{These are all logical-valued functions. If-object (ifo), returns true if it's argument is non-null. If-empty (ife), returns true if it's argument has length zero. If-nonempty (ifn), returns true, if it's argument has length one or greater. Note that the argument name "object", may be changed in future.}
\usage{
ifo (object)
ife (object)
ifn (object)
}
\arguments{
\item{object}{.}
}

