\name{deprecated-functions}
\alias{ifst}
\alias{if0}
\alias{if1}
\alias{iter}
\alias{itobj}
\alias{preview}
\title{deprecated functions}
\description{Do not use.}
\usage{
ifst (obj)
if0 (obj)
if1 (obj)
iter (n, min=1)
itobj (obj)
preview (\dots)
}
\arguments{
\item{obj}{.}
\item{n}{.}
\item{min}{.}
\item{\dots}{.}
}




