\name{enhanced_functions}
\alias{FUNCTION}
\alias{is.FUNCTION}
\alias{s3x_print.FUNCTION}
\alias{s3x_format.FUNCTION}
\alias{$.FUNCTION}
\alias{$<-.FUNCTION}
\title{enhanced functions}
\description{Enhanced functions are R functions with an alternative attribute system. Note that, in the constructor, the \dots argument is ignored.}
\usage{
FUNCTION (f, \dots, xlim, ylim, zlim)
is.FUNCTION (f)
\method{s3x_print}{FUNCTION}(f, ...)
\method{s3x_format}{FUNCTION}(f, ...)
\method{$}{FUNCTION}(f, name)
\method{$}{FUNCTION}(f, name) <- value
}
\arguments{
\item{f}{A function}
\item{xlim}{.}
\item{ylim}{.}
\item{zlim}{.}
\item{name}{.}
\item{value}{.}
\item{\dots}{.}
}

