\name{enhanced_lists}
\alias{LIST}
\alias{s3x_print.LIST}
\alias{s3x_format.LIST}
\alias{[.LIST}
\title{enhanced lists}
\description{Enhanced lists are extended R lists. In themselves, they offer no added functionality. However, enhanced lists have bridge functions, which can simplify subclasses.}
\usage{
LIST (\dots)
\method{s3x_print}{LIST}(object, \dots)
\method{s3x_format}{LIST}(object, \dots)
\method{[}{LIST}(object, \dots)
}
\arguments{
\item{object}{An enhanced list.}
\item{\dots}{.}
}

