\name{enhanced_tables}
\alias{TABLE}
\alias{is_TABLE}
\alias{print.TABLE}
\alias{summary.TABLE}
\alias{plot.TABLE}
\alias{$.TABLE}
\alias{$<-.TABLE}
\alias{[.TABLE}
\alias{[[.TABLE}
\alias{[[<-.TABLE}
\alias{dim.TABLE}
\alias{as.list.TABLE}
\alias{as.data.frame.TABLE}
\alias{as.matrix.TABLE}
\title{enhanced tables}
\description{Refer to the package vignette. Note that enhanced tables are early prototypes.}
\usage{
TABLE (object)
is_TABLE (object)
\method{print}{TABLE}(table, ndp, \dots)
\method{summary}{TABLE}(table, \dots)
\method{plot}{TABLE}(table, \dots)
\method{$}{TABLE}(table, name)
\method{$}{TABLE}(table, name) <- value
\method{[}{TABLE}(table, i, j, \dots)
\method{[[}{TABLE}(table, i, j, \dots)
\method{[[}{TABLE}(table, i, j, \dots) <- value
\method{dim}{TABLE}(x, \dots)
\method{as.list}{TABLE}(table, \dots)
\method{as.data.frame}{TABLE}(table, \dots)
}
\arguments{
\item{object}{.}
\item{table}{.}
\item{ndp}{.}
\item{name}{.}
\item{value}{.}
\item{i}{.}
\item{j}{.}
\item{x}{.}
\item{\dots}{.}
}


