\name{enhanced_vectors}
\alias{VECTOR}
\alias{is.VECTOR}
\alias{s3x_as.data.frame.VECTOR}
\alias{s3x_print.VECTOR}
\alias{$.VECTOR}
\alias{$<-.VECTOR}
\alias{[.VECTOR}
\title{enhanced vectors}
\description{Enhanced vectors are R vectors with an alternative attribute system.}
\usage{
VECTOR (v)
is.VECTOR (v)
\method{s3x_as.data.frame}{VECTOR}(v, \dots)
\method{s3x_print}{VECTOR}(v, \dots)
\method{$}{VECTOR}(v, name)
\method{$}{VECTOR}(v, name) <- value
\method{[}{VECTOR}(v, \dots)
}
\arguments{
\item{v}{A vector.}
\item{name}{.}
\item{value}{.}
\item{\dots}{.}
}

