\name{ENVIRONMENT}
\alias{ENVIRONMENT}
\alias{is.ENVIRONMENT}
\alias{is.hashed}
\alias{==.ENVIRONMENT}
\alias{print.ENVIRONMENT}
\title{enhanced environments}
\description{Enhanced environments, with simplified constructors and support functions.}
\usage{
ENVIRONMENT(\dots, call=sys.call (), hash=FALSE)
is.ENVIRONMENT(e)
is.hashed (e)
\method{==}{ENVIRONMENT}(e1, e2)
\method{print}{ENVIRONMENT}(e, \dots)
}
\arguments{
\item{call}{The call used to resolve argument names.}
\item{hash}{Whether or not the ENVIRONMENT is hashed. Refer to help for new.env for more information.}
\item{e}{An ENVIRONMENT object.}
\item{e1}{.}
\item{e2}{.}
\item{\dots}{The environment's objects.}
}




