\name{general_table_utilities}
\alias{where}
\alias{where.TABLE}
\alias{sort.TABLE}
\alias{sample.TABLE}
\title{general table utilities}
\description{Early prototype functions for subsetting and sorting tables. The sample method calls sample_deterministic.}
\usage{
where (\dots)
\method{where}{TABLE}(table, variable, value, \dots)
\method{sort}{TABLE}(table, by=1, increasing=TRUE, \dots)
\method{sample}{TABLE}(table, \dots)}
\arguments{
\item{table}{.}
\item{variable}{.}
\item{value}{.}
\item{by}{.}
\item{increasing}{.}
\item{\dots}{.}
}




