\name{interpolate}
\alias{interpolate}
\alias{plot.interpolate}
\alias{interpolate_points}
\title{interpolation functions}
\description{The function interpolate, creates a function for linear interpolation of the given series. The function interpolate_points, interpolates between 2 points.}
\usage{
interpolate (x, y)
\method{plot}{interpolate}(f, points=FALSE, \dots)
interpolate_points (x1, y1, x2, y2, u)
}
\arguments{
\item{f}{Interpolation function.}
\item{x}{Series x values.}
\item{y}{Series y values.}
\item{x1}{First point, x value.}
\item{y1}{First point, y value.}
\item{x2}{Second point, x value.}
\item{y2}{Second point, y value.}
\item{u}{Interpolated point, x value}
\item{points}{.}
\item{\dots}{.}
}

