\name{iterators}
\alias{seqv}
\alias{seqv.ol}
\title{iteration utilities}
\description{The function seqv takes an integer n, if n is greater than or equal to one, it returns 1:n, otherwise, it returns NULL. The function seqv.ol, calls seqv using the length of it's argument.}
\usage{
seqv (n)
seqv.ol (object)
}
\arguments{
\item{n}{.}
\item{object}{.}
}


