\name{masks}
\alias{print}
\alias{format}
\alias{summary}
\alias{plot}
\alias{lines}
\alias{points}
\alias{mean}
\title{mask functions for standard generics}
\description{These functions mask standard generics, for the purposes of getting around R check's S3 restrictions.}
\usage{
print (\dots)
format (\dots)
summary (\dots)
plot (\dots)
lines (\dots)
points (\dots)
}
\arguments{
\item{\dots}{The arguments.}
}



