\name{redef_base}
\alias{all.equal}
\alias{anyDuplicated}
\alias{as.array}
\alias{as.data.frame}
\alias{as.Date}
\alias{as.expression}
\alias{as.function}
\alias{as.list}
\alias{as.matrix}
\alias{as.null}
\alias{as.POSIXct}
\alias{as.POSIXlt}
\alias{as.single}
\alias{as.table}
\alias{by}
\alias{chol}
\alias{close}
\alias{conditionCall}
\alias{conditionMessage}
\alias{cut}
\alias{determinant}
\alias{diff}
\alias{duplicated}
\alias{flush}
\alias{format}
\alias{getDLLRegisteredRoutines}
\alias{is.na<-}
\alias{isSymmetric}
\alias{julian}
\alias{kappa}
\alias{labels}
\alias{levels}
\alias{mean}
\alias{merge}
\alias{months}
\alias{open}
\alias{print}
\alias{qr}
\alias{quarters}
\alias{rev}
\alias{row.names}
\alias{row.names<-}
\alias{rowsum}
\alias{scale}
\alias{seek}
\alias{seq}
\alias{solve}
\alias{sort}
\alias{split}
\alias{split<-}
\alias{subset}
\alias{summary}
\alias{t}
\alias{toString}
\alias{transform}
\alias{truncate}
\alias{unique}
\alias{units}
\alias{units<-}
\alias{weekdays}
\alias{with}
\alias{within}
\alias{xtfrm}
\title{redefinitions for base package}
\description{Redefinitions for the base package's generic functions, such that there are no named arguments.}
\usage{
all.equal (\dots)
anyDuplicated (\dots)
as.array (\dots)
as.data.frame (\dots)
as.Date (\dots)
as.expression (\dots)
as.function (\dots)
as.list (\dots)
as.matrix (\dots)
as.null (\dots)
as.POSIXct (\dots)
as.POSIXlt (\dots)
as.single (\dots)
as.table (\dots)
by (\dots)
chol (\dots)
close (\dots)
conditionCall (\dots)
conditionMessage (\dots)
cut (\dots)
determinant (\dots)
diff (\dots)
duplicated (\dots)
flush (\dots)
format (\dots)
getDLLRegisteredRoutines (\dots)
is.na (\dots) <- value
isSymmetric (\dots)
julian (\dots)
kappa (\dots)
labels (\dots)
levels (\dots)
mean (\dots)
merge (\dots)
months (\dots)
open (\dots)
print (\dots)
qr (\dots)
quarters (\dots)
rev (\dots)
row.names (\dots)
row.names (\dots) <- value
rowsum (\dots)
scale (\dots)
seek (\dots)
seq (\dots)
solve (\dots)
sort (\dots)
split (\dots)
split (\dots) <- value
subset (\dots)
summary (\dots)
t (\dots)
toString (\dots)
transform (\dots)
truncate (\dots)
unique (\dots)
units (\dots)
units (\dots) <- value
weekdays (\dots)
with (\dots)
within (\dots)
xtfrm (\dots)
}
\arguments{
\item{\dots}{.}
\item{value}{.}
}




