\name{redef_graphics}
\alias{Axis}
\alias{barplot}
\alias{boxplot}
\alias{cdplot}
\alias{contour}
\alias{hist}
\alias{identify}
\alias{image}
\alias{lines}
\alias{mosaicplot}
\alias{pairs}
\alias{persp}
\alias{plot}
\alias{points}
\alias{spineplot}
\alias{stripchart}
\alias{text}
\title{redefinitions for graphics package}
\description{Redefinitions for the graphics package's generic functions, such that there are no named arguments.}
\usage{
Axis (\dots)
barplot (\dots)
boxplot (\dots)
cdplot (\dots)
contour (\dots)
hist (\dots)
identify (\dots)
image (\dots)
lines (\dots)
mosaicplot (\dots)
pairs (\dots)
persp (\dots)
plot (\dots)
points (\dots)
spineplot (\dots)
stripchart (\dots)
text (\dots)
}
\arguments{
\item{\dots}{.}
}



