\name{s3x_generics}
\alias{s3x_print}
\alias{s3x_format}
\alias{s3x_as.data.frame}
\alias{s3x_plot}
\alias{s3x_lines}
\alias{s3x_points}
\title{s3x generic functions}
\description{The s3x package contains s3x generics, that mimic the standard s3 generics, except that they have no arguments, other than \dots.}
\usage{
s3x_print (\dots)
s3x_format (\dots)
s3x_as.data.frame (\dots)
s3x_plot (\dots)
s3x_lines (\dots)
s3x_points (\dots)
}
\arguments{
\item{\dots}{.}
}

