\name{vector_style}
\alias{vector_style}
\alias{print.vector_style}
\alias{format.vector_style}
\title{formatted vectors}
\description{An object for representing integer or numeric vectors, with formatting, including currency symbols and thousands separators.}
\usage{
vector_style (x, ndp=2,
    currency.symbol="", decimal.sep=".", thousands.sep =",", fill=TRUE)
\method{print}{vector_style}(x, \dots)
\method{format}{vector_style}(x, \dots)
}
\arguments{
\item{x}{.}
\item{ndp}{.}
\item{currency.symbol}{.}
\item{decimal.sep}{.}
\item{thousands.sep}{.}
\item{fill}{.}
\item{\dots}{.}
}


