\name{enhanced_functions}
\alias{FUNCTION}
\alias{is_FUNCTION}
\alias{print.FUNCTION}
\alias{plot.FUNCTION}
\alias{lines.FUNCTION}
\alias{format.FUNCTION}
\alias{seq.FUNCTION}
\alias{$.FUNCTION}
\alias{$<-.FUNCTION}
\title{enhanced functions}
\description{Refer to the package vignette.}
\usage{
FUNCTION (f, \dots, xlim=NULL, ylim=NULL, zlim=NULL)
is_FUNCTION (object)
\method{print}{FUNCTION}(f, \dots)
\method{plot}{FUNCTION}(f, \dots, xlim, ylim, xlab="x", ylab="f(x)", n=200)
\method{lines}{FUNCTION}(f, \dots, xlim, n)
\method{format}{FUNCTION}(f, \dots)
\method{seq}{FUNCTION}(f, \dots, xlim=f$xlim, n)
\method{$}{FUNCTION}(f, name)
\method{$}{FUNCTION}(f, name) <- value
}
\arguments{
\item{f}{.}
\item{object}{.}
\item{xlim}{.}
\item{ylim}{.}
\item{zlim}{.}
\item{xlab}{.}
\item{ylab}{.}
\item{n}{.}
\item{name}{.}
\item{value}{.}
\item{\dots}{.}
}

