\name{enhanced_vectors}
\alias{VECTOR}
\alias{is_VECTOR}
\alias{print.VECTOR}
\alias{as.data.frame.VECTOR}
\alias{$.VECTOR}
\alias{$<-.VECTOR}
\alias{[.VECTOR}
\alias{[[.VECTOR}
\title{enhanced vectors}
\description{Refer to the package vignette.}
\usage{
VECTOR (v)
is_VECTOR (object)
\method{print}{VECTOR}(v, \dots)
\method{as.data.frame}{VECTOR}(v, \dots)
\method{$}{VECTOR}(v, name)
\method{$}{VECTOR}(v, name) <- value
\method{[}{VECTOR}(v, \dots)
\method{[[}{VECTOR}(v, \dots)
}
\arguments{
\item{v}{.}
\item{object}{.}
\item{name}{.}
\item{value}{.}
\item{\dots}{.}
}


